Consolidate test classes, move MockServer, FileSystemBackedV3MockServer and StreamExtensions to Test.Utility (#4797)
