+++ test/NuGet.Clients.Tests/NuGet.CommandLine.Test/FeedPackagePruningTests.cs
@@ -10,6 +10,7 @@ using Newtonsoft.Json.Linq;
 using NuGet.Frameworks;
 using NuGet.Protocol;
 using NuGet.Test.Utility;
+using Test.Utility;
 using Xunit;
 
 namespace NuGet.CommandLine.Test
