+++ test/NuGet.Core.FuncTests/Dotnet.Integration.Test/SignCommandTestFixture.cs
@@ -10,6 +10,7 @@ using System.Security.Cryptography.X509Certificates;
 using System.Threading.Tasks;
 using NuGet.Packaging.Signing;
 using NuGet.Test.Utility;
+using Test.Utility;
 using Test.Utility.Signing;
 
 namespace Dotnet.Integration.Test
