+++ test/TestUtilities/Test.Utility/FileSystemBackedV3MockServer.cs
@@ -10,11 +10,10 @@ using Newtonsoft.Json.Linq;
 using NuGet.Common;
 using NuGet.Packaging.Core;
 using NuGet.Protocol;
-using Test.Utility;
 
-namespace Dotnet.Integration.Test
+namespace Test.Utility
 {
-    internal class FileSystemBackedV3MockServer : MockServer
+    public class FileSystemBackedV3MockServer : MockServer
     {
         private string _packageDirectory;
         private readonly MockResponseBuilder _builder;
@@ -87,7 +86,7 @@ namespace Dotnet.Integration.Test
                             response.ContentType = "application/zip";
                             using (var stream = file.OpenRead())
                             {
-                                var content = ReadAllBytes(stream);
+                                var content = stream.ReadAllBytes();
                                 SetResponseContent(response, content);
                             }
                         });
@@ -118,7 +117,7 @@ namespace Dotnet.Integration.Test
                         {
                             response.ContentType = "text/javascript";
                             var packageToListedMapping = packages.Select(e => new KeyValuePair<PackageIdentity, bool>(e.Identity, !UnlistedPackages.Contains(e.Identity))).ToArray();
-                            MockResponse mockResponse = _builder.BuildRegistrationIndexResponse(ServiceIndexUri, packageToListedMapping);
+                            MockResponse mockResponse = _builder.BuildRegistrationIndexResponse(Uri, packageToListedMapping);
                             SetResponseContent(response, mockResponse.Content);
                         });
                     }
@@ -139,23 +138,6 @@ namespace Dotnet.Integration.Test
             {
                 // Debug here
                 throw;
-            }
-        }
-
-        public static byte[] ReadAllBytes(Stream stream)
-        {
-            var memoryStream = stream as MemoryStream;
-            if (memoryStream != null)
-            {
-                return memoryStream.ToArray();
-            }
-            else
-            {
-                using (memoryStream = new MemoryStream())
-                {
-                    stream.CopyTo(memoryStream);
-                    return memoryStream.ToArray();
-                }
             }
         }
     }
